changed types according to the last changes in documentation

+++ src/main/java/ru/cft/shift/intensive/template/dto/PurchaseDto.java
@@ -8,7 +8,7 @@ public class PurchaseDto {
     @NotEmpty private String orderId;
     @NotEmpty private Address shopAddress;
     @NotEmpty private PurchaseState state;
-    @NotEmpty @Min(value = 1) private Float weight;
+    @NotEmpty @Min(value = 1) private Integer weight;
     @NotEmpty private String shopName;
     @NotEmpty private String itemId;
     @NotEmpty private String itemName;
@@ -20,7 +20,7 @@ public class PurchaseDto {
     public PurchaseDto() {
     }
 
-    public PurchaseDto(String orderId, String shopAddress, String state, Float weight, String shopName, 
+    public PurchaseDto(String orderId, String shopAddress, String state, Integer weight, String shopName, 
     String itemId, String itemName, Integer quantity, Float cost, String customerAddress, String description) throws RuntimeException {
         Boolean isInputValid = true;
         for (Address district: Address.values()) {
@@ -83,11 +83,11 @@ public class PurchaseDto {
         this.state = state;
     }
 
-    public Float getWeight() {
+    public Integer getWeight() {
         return weight;
     }
 
-    public void setWeight(Float weight) {
+    public void setWeight(Integer weight) {
         this.weight = weight;
     }
